+++ app/src/main/java/com/adammcneilly/spacenerd/core/displaymodels/LaunchDisplayModel.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.spacenerd.core.displaymodels
 
-import com.adammcneilly.spacenerd.core.models.Launch
-import com.adammcneilly.spacenerd.core.models.LaunchStatus
+import com.adammcneilly.spacenerd.shared.core.models.Launch
+import com.adammcneilly.spacenerd.shared.core.models.LaunchStatus
 
 data class LaunchDisplayModel(
     val id: String,
