--- app/src/main/java/com/adammcneilly/spacenerd/core/models/Author.kt
@@ -1,5 +1,0 @@
-package com.adammcneilly.spacenerd.core.models
-
-data class Author(
-    val name: String,
-)
