+++ app/src/main/java/com/adammcneilly/spacenerd/data/repositories/ArticleRepository.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.spacenerd.data.repositories
 
-import com.adammcneilly.spacenerd.core.models.Article
+import com.adammcneilly.spacenerd.shared.core.models.Article
 import kotlinx.coroutines.flow.Flow
 
 interface ArticleRepository {
