+++ app/src/main/java/com/adammcneilly/spacenerd/data/repositories/LaunchRepository.kt
@@ -1,7 +1,7 @@
 package com.adammcneilly.spacenerd.data.repositories
 
-import com.adammcneilly.spacenerd.core.models.Launch
 import com.adammcneilly.spacenerd.data.models.LaunchListRequest
+import com.adammcneilly.spacenerd.shared.core.models.Launch
 import kotlinx.coroutines.flow.Flow
 
 interface LaunchRepository {
