+++ app/src/main/java/com/adammcneilly/spacenerd/data/thespacedevs/dto/TSDPadDTO.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.spacenerd.data.thespacedevs.dto
 
-import com.adammcneilly.spacenerd.core.models.LaunchPad
+import com.adammcneilly.spacenerd.shared.core.models.LaunchPad
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
