+++ app/src/main/java/com/adammcneilly/spacenerd/data/thespacedevs/dto/TSDStatusDTO.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.spacenerd.data.thespacedevs.dto
 
-import com.adammcneilly.spacenerd.core.models.LaunchStatus
+import com.adammcneilly.spacenerd.shared.core.models.LaunchStatus
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
