+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/core/models/Author.kt
@@ -1,0 +1,5 @@
+package com.adammcneilly.spacenerd.shared.core.models
+
+data class Author(
+    val name: String,
+)
