+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/core/models/Launch.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.spacenerd.core.models
+package com.adammcneilly.spacenerd.shared.core.models
 
 import kotlinx.datetime.Instant
