+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/core/models/LaunchServiceProvider.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.spacenerd.core.models
+package com.adammcneilly.spacenerd.shared.core.models
 
 data class LaunchServiceProvider(
     val id: String,
